pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = "==3.12.*"
dependencies = [
    "air-drf-relation>=0.6.1",
    "celery>=5.4.0",
    "django>=5.1.2",
    "django-cors-headers>=4.5.0",
    "django-environ>=0.11.2",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.2",
    "psycopg[binary]>=3.2.3",
    "redis>=5.2.0",
    "requests>=2.32.3",
    "sentry-sdk>=2.17.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.1",
]

[tool.pyupgrade]
py312-plus = true

[tool.ruff]
fix = true
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.extend-per-file-ignores]
"core/settings/*.py" = ["F405", "F403"]

[tool.ruff.format]
quote-style = "single"
skip-magic-trailing-comma = false
line-ending = "auto"
